Split co-ai deploy packaging helpers
